move soft delete into prisma extension
